cambios daos pruebas

+++ java/vo/Sala.java
@@ -1,4 +1,4 @@
 package vo;
 public class Sala {
-		public Long n;
+		public Integer n;
 }
